requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.30
alembic==1.13.1

python-dotenv==1.0.1
pydantic==2.7.3
psycopg2-binary==2.9.9

# Optional tools
httpx==0.27.0            # Jika perlu request async
passlib[bcrypt]==1.7.4   # Jika pakai hashing password

cryptography

python-multipart
uvicorn
twilio
gspread 
oauth2client
pypdf 
faiss-cpu 
openai
langchain-community
sentence-transformers